Various searching features (Expand for details):
feat: Added searchView in fragment. Ensured it works.
feat: Now an item matches on lowercase strings.
feat: Now an item matches on either name or address.
feat: Added Markable to indicate an Object that can be marked based on some criteria. In the case of LogItems a SpannableString version of name and address has been added, and the substring matching the query will be highlighted.