feat: implemented the search option in the contact list
User writes a text in the SearchView of the ContactListActivity. Text could be part of a name of a contact or his address. Research isn't case sensitive.
The ContactSearchListener actives onQueryTextChange method that invokes the ContactFilter.
ContactFilter takes care of filtering the data then passes them to the adapter that displays them.