feat: define ContactListClickListener and a ViewModel to share selected contact between fragments
ContactListClickListener is used to define a schema invoked on the click on a contact item.
After a OnClick on the item, the SMSContact selected is passed on ContactListClickListener#OnClick.
The implementation of this method is in the ContactListFragment because it requires some instructions accessible only inside a Fragment.
In this implementation the contact is saved in the ContactSharedViewModel.
When the HomeFragment is resumed, the contact is got from ContactSharedViewModel#getSelected().